requirements.txt
# === CORE LLM + LANGCHAIN ===
langchain==1.0.7
langchain-core==1.0.5
langchain-mistralai==1.0.1
langgraph==1.0.3
langgraph-sdk==0.2.9
langsmith==0.4.43
langgraph-checkpoint-postgres==3.0.4
pydantic-settings==2.13.1
thefuzz==0.22.1

# === MISTRAL ===
httpx==0.28.1
tenacity==9.1.2
python-dotenv==1.0.1

# === FINANCE DATA ===
yfinance==1.2.0
pandas==2.2.3
numpy==2.4.3
pyportfolioopt==1.6.0
dacite==1.6.0

# === yfinance dependencies ===
beautifulsoup4==4.12.3
platformdirs==4.3.6
multitasking==0.0.12
frozendict==2.4.7
peewee==3.18.3
curl_cffi==0.13.0
#websockets==12.0
protobuf==4.25.3

# === UTILITIES ===
requests==2.32.3
typing_extensions==4.15.0
cffi>=1.17.0

# === FRONTEND === 
fastapi==0.124.0
uvicorn==0.38.0

# === DATA BASE ===
argon2-cffi==25.1.0
cryptography==46.0.3
supabase==2.25.1
psycopg_pool==3.3.0
psycopg-binary==3.3.3

# === ML / INTENT CLASSIFIER ===
sentence-transformers==3.0.1
scikit-learn==1.5.2
joblib==1.4.2
torch==2.10.0
tqdm==4.66.5

# === TESTS ===
pytest==9.0.2
pytest-asyncio==1.3.0